// src/models/usermodel.ts
export class usermodel {
    CustName: string;
    MobileNo: string;
    Email: string;
    Product: string;
    SubProduct: string;
    Module: string;
    Deal: string;
    riskPincode: string;
    StateCode: number;
    StateTxt: string;
    CityCode: number;
    CityTxt: string;
    Address: string;
}